Extract shared webpack patcher helper, remove duplicated patching code
Move the webpack module patching boilerplate (used to override Next.js
bundled exports in E2E tests) into a shared registerWebpackPatcher()
helper in xstate-mocks.ts. This removes ~280 lines of copy-pasted code
across ban-enforcement, blackout-periods, edit-booking-flow, and
safety-training tests. Also removes unused import in form-validation.
